~ UserServer.java, + threadedConnection.java
Changed UserServer.java, added threadedConnection.java
The server instance will now create a new thread when there's an incoming connection, and send that connection to the new thread.

threadedConnection does all the heavy lifting when it comes to handling the requests of the incoming connection, and responding with the appropriate data, and/or error code.

Also removed the deprecated amazon authentication method, and replaced it with the up to date supported method.
